backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.10.0
pydantic==2.5.3
matplotlib==3.8.2
numpy==1.26.3
